+++ lib/handlers/get-limit-orders/index.ts
@@ -10,7 +10,7 @@ import { ONE_DAY_IN_SECONDS } from '../../util/constants'
 import { log } from '../../Logging'
 import { OrderDispatcher } from '../../services/OrderDispatcher'
 import { OffChainRelayOrderValidator } from '../../util/OffChainRelayOrderValidator'
-import { OffChainUniswapXOrderValidator } from '../../util/order-validator'
+import { OffChainUniswapXOrderValidator } from '../../util/OffChainUniswapXOrderValidator'
 import { GetOrdersHandler } from '../get-orders/handler'
 import { OnChainValidatorMap } from '../OnChainValidatorMap'
 import { getMaxOpenOrders } from '../post-order/injector'
